+++ NEWS.md
@@ -1,8 +1,13 @@
 # Coming Soon
 
 * Geo-fencing.
-* Modify to accept MySQL connections
 * Open sentiment analysis function so that sentiment analysis can be done post collection.
+
+# twinfoR 1.0.2
+
+* Works with MySQL connections.
+* Fixed minor bugs.
+* Added open_status function.
 
 # twinfoR 1.0.1